requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
fonttools==4.33.3
geographiclib==1.52
geojson==2.5.0
geojsonio==0.0.3
geopy==2.2.0
github3.py==3.2.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
PyMySql==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.9
